+++ CHANGES.md
@@ -1,4 +1,4 @@
-### HEAD
+### 1.0.1 / 2015-01-02
 
   * Fixed a bug with using Last-Modified and ETag headers with mach.file
   * Added support for binary response bodies in IE9